+++ packages/server/src/index.ts
@@ -1,6 +1,6 @@
 import server from 'common/server';
 import WebsocketServer from 'ws/server';
-import config from 'config';
+import config from 'config/config';
 
 function startServer() {
     const port = parseInt(config.port);
