+++ packages/server/src/util/redis.ts
@@ -1,4 +1,4 @@
-import config from 'config';
+import config from 'config/config';
 import Redis from 'ioredis';
 
 let redis: Redis.Redis;
